// src/app/shared/interfaces/Character/character.ts
export interface Character {
  birth_year: string;
  eye_color: string;
  filmIds: number[];
  films: string[];
  gender: string;
  hair_color: string;
  height: string;
  homeworld: string;
  homeworldId: number;
  id: number;
  mass: string;
  name: string;
  skin_color: string;
  speciesIds: number[];
  species: string[];
  starshipIds: number[];
  starships: string[];
  url: string;
  vehicleIds: number[];
  vehicles: string[];
}